모바일 디텍터 추가
