+++ front/package.json
@@ -17,6 +17,7 @@
     "chart.js": "^2.9.3",
     "jquery": "^3.5.1",
     "lightweight-charts": "^2.0.0",
+    "mobile-detect": "^1.4.4",
     "next": "^9.3.2",
     "next-redux-wrapper": "^3.0.0-alpha.2",
     "node-sass": "^4.14.1",
@@ -25,6 +26,7 @@
     "react": "^16.8.6",
     "react-burger-menu": "^2.6.13",
     "react-chartjs-2": "^2.9.0",
+    "react-device-detect": "^1.12.1",
     "react-dom": "^16.8.6",
     "react-redux": "^7.1.0-alpha.4",
     "redux": "^4.0.1",
